Insturmentation key can be overriden by code, TelemetryClient interface and concrete moved
